requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
opencv-python-headless
python-multipart
Pillow
huggingface-hub

